moved docsbutton to own file, added copy markdown button
