+++ frontend/src/lib/features/prototypes/components/ShowPrototypes.tsx
@@ -4,9 +4,10 @@ import { useSystemPrototypes } from "$lib/features/prototypes/queries";
 import { prototypeURL } from "$shared/globals";
 import { Button, CircularProgress, Divider, Modal, ModalClose, ModalDialog } from '@mui/joy';
 import { useQueryClient } from "@tanstack/react-query";
-import { BookText, FileText, Loader2, Package, Play, Trash } from "lucide-react";
+import { FileText, Package, Play, Trash } from "lucide-react";
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router";
+import { DocsButton } from "./DocsButton";
 
 
 type Props = {
@@ -20,9 +21,7 @@ export const ShowPrototypes: React.FC<Props> = () => {
     const [loading, setLoading] = useState<{ [key: string]: boolean }>({});
     const [metadata, setMetadata] = useState("");
     const [showMetadataModal, setShowMetadataModal] = useState(false);
-    const [showDocsModal, setShowDocsModal] = useState(false);
-    const [docs, setDocs] = useState("");
-    const [docsLoading, setDocsLoading] = useState(false);
+
 
     const queryClient = useQueryClient();
     useEffect(() => {
@@ -125,50 +124,14 @@ export const ShowPrototypes: React.FC<Props> = () => {
         }
     }
 
-    const showDocs = async (prototypeId: string) => {
-        setDocsLoading(true);
-        await new Promise((resolve) => setTimeout(resolve, 2000));
 
-        setDocs(`
-# UML Prototype Documentation
-
-## Table of Contents
-1. [Introduction](#introduction)
-2. [Project Overview](#project-overview)
-3. [Design Goals](#design-goals)
-4. [UML Diagrams](#uml-diagrams)
-   - [Class Diagram](#class-diagram)
-   - [Sequence Diagram](#sequence-diagram)
-   - [Use Case Diagram](#use-case-diagram)
-   - [Activity Diagram](#activity-diagram)
-5. [Component Descriptions](#component-descriptions)
-6. [Assumptions and Constraints](#assumptions-and-constraints)
-7. [Future Enhancements](#future-enhancements)
-8. [References](#references)
-
----
-
-## Introduction
-
-This document provides an overview and detailed description of the UML prototype. It is intended to help developers, designers, and stakeholders understand the system architecture, relationships, and interactions.
-
-
-            `);
-        setShowDocsModal(true);
-        setDocsLoading(false);
-
-
-    }
 
     const closeMetadataModal = () => {
         setMetadata("");
         setShowMetadataModal(false);
     }
 
-    const closeDocs = () => {
-        setDocs("");
-        setShowDocsModal(false);
-    }
+
 
     return (
         <>
@@ -217,13 +180,7 @@ This document provides an overview and detailed description of the UML prototype
                                 </td>
                                 <td className="py-2 px-4 text-left border-b border-gray-200  flex space-x-2">
 
-                                <button
-                                        onClick={() => showDocs(e.id)}
-                                        className="w-[80px] h-[40px] bg-stone-200 rounded-md hover:bg-stone-300 flex items-center justify-center shrink-0"
-                                    >
-                                     {docsLoading ? <Loader2 className="animate-spin size-4 mr-2 shrink-0" /> : <BookText className="size-4 mr-2 shrink-0" />}
-                                        Docs
-                                    </button>
+                               <DocsButton />
 
                                     <button
                                         onClick={() => showMetadata(e.id)}
@@ -311,21 +268,7 @@ This document provides an overview and detailed description of the UML prototype
                     </div>
                 </ModalDialog>
             </Modal>
-            <Modal
-                open={showDocsModal}
-                onClose={closeDocs}
-            >
-                <ModalDialog className="max-h-screen overflow-y-auto">
-                    <ModalClose
-                        sx={{
-                            position: "relative",
-                        }}
-                    />
-                    <div className="flex h-full w-full flex-col gap-1 p-3">
-                        <pre>{docs}</pre>
-                    </div>
-                </ModalDialog>
-            </Modal>
+
         </>
     );
 };
